Move most error logging code to a new file in telemetry-utils
