+++ packages/loader/container-utils/src/error.ts
@@ -10,7 +10,12 @@ import {
     IErrorBase,
     IThrottlingWarning,
 } from "@fluidframework/container-definitions";
-import { isILoggingError, annotateError, LoggingError } from "@fluidframework/telemetry-utils";
+import {
+    isILoggingError,
+    annotateError,
+    extractLogSafeErrorProperties,
+    LoggingError,
+} from "@fluidframework/telemetry-utils";
 import { ITelemetryProperties } from "@fluidframework/common-definitions";
 import { ISequencedDocumentMessage } from "@fluidframework/protocol-definitions";
 
@@ -18,44 +23,6 @@ import { ISequencedDocumentMessage } from "@fluidframework/protocol-definitions"
 const hasErrorType = (error: any): error is IErrorBase => {
     return (typeof error?.errorType === "string");
 };
-
-const isRegularObject = (value: any): boolean => {
-    return value !== null && !Array.isArray(value) && typeof value === "object";
-};
-
-function extractLogSafeErrorProperties(error: any) {
-    const removeMessageFromStack = (stack: string, errorName?: string) => {
-        const stackFrames = stack.split("\n");
-        stackFrames.shift(); // Remove "[ErrorName]: [ErrorMessage]"
-        if (errorName !== undefined) {
-            stackFrames.unshift(errorName); // Add "[ErrorName]"
-        }
-        return stackFrames.join("\n");
-    };
-
-    const message = (typeof error?.message === "string")
-        ? error.message as string
-        : String(error);
-
-    const safeProps: { message: string; errorType?: string; stack?: string } = {
-        message,
-    };
-
-    if (isRegularObject(error)) {
-        const { errorType, stack, name } = error;
-
-        if (typeof errorType === "string") {
-            safeProps.errorType = errorType;
-        }
-
-        if (typeof stack === "string") {
-            const errorName = (typeof name === "string") ? name : undefined;
-            safeProps.stack = removeMessageFromStack(stack, errorName);
-        }
-    }
-
-    return safeProps;
-}
 
 /**
  * Generic wrapper for an unrecognized/uncategorized error object
