+++ packages/utils/telemetry-utils/src/errorLogging.ts
@@ -1,0 +1,172 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import {
+    ILoggingError,
+    ITaggedTelemetryPropertyType,
+    ITelemetryProperties,
+} from "@fluidframework/common-definitions";
+
+/** @returns true if value is an object but neither null nor an array */
+const isRegularObject = (value: any): boolean => {
+    return value !== null && !Array.isArray(value) && typeof value === "object";
+};
+
+/** Inspect the given error for common "safe" props and return them */
+export function extractLogSafeErrorProperties(error: any) {
+    const removeMessageFromStack = (stack: string, errorName?: string) => {
+        const stackFrames = stack.split("\n");
+        stackFrames.shift(); // Remove "[ErrorName]: [ErrorMessage]"
+        if (errorName !== undefined) {
+            stackFrames.unshift(errorName); // Add "[ErrorName]"
+        }
+        return stackFrames.join("\n");
+    };
+
+    const message = (typeof error?.message === "string")
+        ? error.message as string
+        : String(error);
+
+    const safeProps: { message: string; errorType?: string; stack?: string } = {
+        message,
+    };
+
+    if (isRegularObject(error)) {
+        const { errorType, stack, name } = error;
+
+        if (typeof errorType === "string") {
+            safeProps.errorType = errorType;
+        }
+
+        if (typeof stack === "string") {
+            const errorName = (typeof name === "string") ? name : undefined;
+            safeProps.stack = removeMessageFromStack(stack, errorName);
+        }
+    }
+
+    return safeProps;
+}
+
+/** Type guard for ILoggingError */
+export const isILoggingError = (x: any): x is ILoggingError => typeof x?.getTelemetryProperties === "function";
+
+/**
+ * Read-Write Logging Error.  Not exported.
+ * This type alias includes addTelemetryProperties, and applies to objects even if not instanceof LoggingError\
+ */
+type RwLoggingError = LoggingError;
+
+/** Type guard for RwLoggingError.  Not exported. */
+const isRwLoggingError = (x: any): x is RwLoggingError =>
+    typeof x?.addTelemetryProperties === "function" && isILoggingError(x);
+
+/**
+ * Annotate the given error object with the given logging props
+ * @returns The same error object passed in if possible, with telemetry props functionality mixed in
+ */
+export function annotateError(
+    error: unknown,
+    props: ITelemetryProperties,
+): ILoggingError {
+    if (isRwLoggingError(error)) {
+        error.addTelemetryProperties(props);
+        return error;
+    }
+
+    if (isRegularObject(error)) {
+        // Even though it's not exposed, fully implement IRwLoggingError for subsequent calls to annotateError
+        const loggingError = error as RwLoggingError;
+
+        const propsForError = {...props};
+        loggingError.getTelemetryProperties = () => propsForError;
+        loggingError.addTelemetryProperties =
+            (newProps: ITelemetryProperties) => { copyProps(propsForError, newProps); };
+        return loggingError;
+    }
+
+    const message = String(error);
+    return new LoggingError(message, props);
+}
+
+/** Copy props from source onto target, overwriting any keys that are already set on target */
+function copyProps(target: unknown, source: ITelemetryProperties) {
+    Object.assign(target, source);
+}
+
+/**
+ * Type guard to identify if a particular value (loosely) appears to be a tagged telemetry property
+ */
+ export function isTaggedTelemetryPropertyValue(x: any): x is ITaggedTelemetryPropertyType {
+    return (typeof(x?.value) !== "object" && typeof(x?.tag) === "string");
+}
+
+/**
+ * Walk an object's enumerable properties to find those fit for telemetry.
+ */
+function getValidTelemetryProps(obj: any): ITelemetryProperties {
+    const props: ITelemetryProperties = {};
+    for (const key of Object.keys(obj)) {
+        const val = obj[key];
+        switch (typeof val) {
+            case "string":
+            case "number":
+            case "boolean":
+            case "undefined":
+                props[key] = val;
+                break;
+            default: {
+                if (isTaggedTelemetryPropertyValue(val)) {
+                    props[key] = val;
+                } else {
+                    // We don't support logging arbitrary objects
+                    props[key] = "REDACTED (arbitrary object)";
+                }
+                break;
+            }
+        }
+    }
+    return props;
+}
+
+/**
+ * Helper class for error tracking that can be used to log an error in telemetry.
+ * The props passed in (and any set directly on the object after the fact) will be
+ * logged in accordance with the given tag, if present.
+ *
+ * PLEASE take care to properly tag properties set on this object
+ */
+export class LoggingError extends Error implements ILoggingError {
+    constructor(
+        message: string,
+        props?: ITelemetryProperties,
+    ) {
+        super(message);
+        if (props) {
+            this.addTelemetryProperties(props);
+        }
+    }
+
+    /**
+     * Add additional properties to be logged
+     */
+    public addTelemetryProperties(props: ITelemetryProperties) {
+        copyProps(this, props);
+    }
+
+    /**
+     * Get all properties fit to be logged to telemetry for this error
+     */
+    public getTelemetryProperties(): ITelemetryProperties {
+        const taggableProps = getValidTelemetryProps(this);
+        // Include non-enumerable props inherited from Error that would not be returned by getValidTelemetryProps
+        // But if any were overwritten (e.g. with a tagged property), then use the result from getValidTelemetryProps.
+        // Not including the 'name' property because it's likely always "Error"
+        return  {
+            stack: this.stack,
+            message: this.message,
+            ...taggableProps,
+        };
+    }
+}
