+++ packages/utils/telemetry-utils/src/index.ts
@@ -3,6 +3,7 @@
  * Licensed under the MIT License.
  */
 export * from "./debugLogger";
+export * from "./errorLogging";
 export * from "./eventEmitterWithErrorHandling";
 export * from "./events";
 export * from "./logger";
